Revert "Bind MaxChangeSizeTimes max value with ChangeSizeCombo"
This reverts commit 896413d1a3e58a5fdcc219f528552a2172aceb35.

+++ osu.Game/Rulesets/Mods/ModFlashlight.cs
@@ -130,7 +130,7 @@ namespace osu.Game.Rulesets.Mods
             private readonly float changeSizeDecreaseRatio;
             private readonly bool comboBasedSize;
 
-            private readonly float maxChangeSizeTimes;
+            private readonly int maxChangeSizeTimes;
             private readonly float changeSizeCombo;
 
             protected Flashlight(ModFlashlight modFlashlight)
